requirements.txt
accelerate==0.17.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bbox==0.9.4
bitsandbytes==0.36.0.post2
blessed==1.20.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
cmake==3.25.0
diffusers==0.13.1
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
einops-exts==0.0.4
exceptiongroup==1.2.0
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.1
hydra-core==1.3.2
idna==3.6
imageio==2.31.1
importlib-metadata==7.0.1
Jinja2==3.1.2
joblib==1.3.2
lightning-utilities==0.10.0
lit==15.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
networkx==2.8.7
nltk==3.8.1
numpy==1.23.0
nvidia-ml-py==12.535.133
omegaconf==2.3.0
openai==1.14.0
openai-clip==1.0.1
opencv-python==4.6.0.66
overrides==7.3.1
pandas==1.4.0
pathtools==0.1.2
protobuf==4.25.1
psutil==5.9.6
pycairo==1.25.1
pydantic==2.6.4
pydantic_core==2.16.3
pyemd==1.0.0
pyquaternion==0.9.9
pytorch-lightning==2.0.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.0
sentry-sdk==1.38.0
setproctitle==1.3.3
smmap==5.0.1
sniffio==1.3.1
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1+cu118
torch-geometric==2.3.1
torchmetrics==0.11.4
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.27.4
triton==2.0.0
urllib3==2.1.0
wandb==0.15.4
wcwidth==0.2.12
yarl==1.9.3
zstandard==0.22.0
